contiki

+++ virtual-sensor/virtual-sensor.h
@@ -1,7 +1,8 @@
 #ifndef VIRTUAL_SENSOR_H
 #define VIRTUAL_SENSOR_H
 
-float read_temperature();
-float read_humidity();
+int read_temperature();
+
+int read_humidity();
 
 #endif
